requirements.txt
requests
PyYAML
pandas
scikit-learn
matplotlib